requirements.txt
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0
google-genai>=1.0.0
python-dateutil>=2.8.0
pydantic>=2.0.0
numpy>=1.24.0
